Activities

+++ src/main/java/devsupeior/SistemaEventos/entities/Participant.java
@@ -9,7 +9,7 @@ public class Participant {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String name;
 
@@ -20,7 +20,7 @@ public class Participant {
 
     }
 
-    public Participant(Long id, String name, String email) {
+    public Participant(Integer id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -34,11 +34,11 @@ public class Participant {
         this.email = email;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
